requirements.txt
# AI / Agent frameworks
openai>=1.75.0
openai-agents>=0.0.12
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Vector store (local RAG)
chromadb>=0.5.0

# HTTP client for MCP tools
httpx>=0.27.0

# MCP protocol
mcp>=1.0.0

# Config & environment
python-dotenv>=1.0.0

# Redis (memory + Celery broker)
redis>=5.0.0

# Task scheduling
celery>=5.3.0
celery[redis]>=5.3.0

# PDF generation (SecretariatAgent)
reportlab>=4.0.0
weasyprint>=62.0

# Observability (LangSmith tracing)
langsmith>=0.1.0

# Utilities
tenacity>=8.2.0
pydantic>=2.0.0

# Orchestrator HTTP bridge (orchestrator_server.py + dashboard AGENTS_API_URL)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
